refactor: remove barrel files (index.ts) pattern
Replace barrel file re-exports with direct imports to reduce maintenance
overhead. Each import now explicitly references the source module.

Changes:
- src/index.ts: direct imports from specific modules
- src/cli/index.ts: direct imports from generator modules
- src/generator/common.ts: direct imports from adapter modules
- Deleted barrel files: parser/, formatter/, generator/, adapter/, test-utils/
