--- src/formatter/index.ts
@@ -1,9 +1,0 @@
-// Types
-export type { OutputFormatter, FormatterOptions } from "./types";
-export type { MermaidFormatterOptions } from "./mermaid";
-export type { MarkdownFormatterOptions } from "./markdown";
-
-// Implementations
-export { DbmlFormatter } from "./dbml";
-export { MarkdownFormatter } from "./markdown";
-export { MermaidErDiagramFormatter } from "./mermaid";
