--- src/generator/index.ts
@@ -1,16 +1,0 @@
-/**
- * DBML Generator Module
- *
- * Provides functions to generate DBML from Drizzle ORM schema definitions.
- * Supports PostgreSQL, MySQL, and SQLite dialects.
- *
- * JSDoc comments can be extracted from source files and included as DBML Note clauses.
- * Use the `source` option to specify the schema source file or directory, or pass pre-extracted
- * comments via the `comments` option.
- */
-
-export { pgGenerate, PgGenerator } from "./pg";
-export { mysqlGenerate, MySqlGenerator } from "./mysql";
-export { sqliteGenerate, SqliteGenerator } from "./sqlite";
-export { BaseGenerator, writeDbmlFile } from "./common";
-export { DbmlBuilder } from "../formatter/dbml-builder";
